Estimate chunk size on compaction looking at data uncompressed size (#9099)
**What this PR does / why we need it**:
At `compactedIndex.ToIndexFile` we were not computing the size of the
chunk correctly. We were looking at the size of the chunk ref. Since the
size of the chunk ref is always a bunk of bytes, the chunk meta KB field
would always be 0.

https://github.com/grafana/loki/blob/4e893a0a88b2c8c3ea9d1d8b1037f54d8fb7f1af/pkg/storage/stores/tsdb/compactor.go#L367

Instead, we should get the approx KBs by looking at the chunks data
uncompressed size. As we do at `store.IndexChunk`:

https://github.com/grafana/loki/blob/d405162516be78beb3d0712cfd57256c7873fe78/pkg/storage/stores/tsdb/store.go#L189