promtail: support configurable polling interval (#8874)
Currently, Promtail always polls files every 250ms. This PR allows the
polling period to be customized, starting at a minimum polling period
and exponentially increasing the polling frequency to a bounded maximum
if the file has no changes.

When file changes are detected, the polling frequency resets to the
minimum.

The default behavior is the existing behavior; both the minimum and
maximum polling intervals are set to 250ms.

I'm opening this in draft to get early feedback. There's a few open
questions:

* Where can I add validations to ensure that the min/max polling
intervals are configured properly?
* Is this the right approach for implementation? Do I need to make any
changes?
* Does this need tests? 

I'll update the CHANGELOG and Documentation after we're agreed on an
approach.