Fix parsing of long queries (#9134)
Loki uses a `scanner.Scanner` to consume the characters in a query.
`lex.go` relies heavily on the assumption a scanner can be copied to
perform speculative parsing. The problem is that sometimes speculative
parsing overruns the scanners internal buffer of `1024` bytes and incurs
an additional read to the underlying reader. At that point, all
subsequent scanners are out of sync with their underlying readers and
the query may fail to be parsed.

In the case of #9132, there is a function call at line `1024`. `lex.go`
tries to understand this is a function call by speculatively parsing and
incurs an underlying read which consumers the rest of the reader. When
the lexer tries to scan the next character with the original scanner,
the reader is already consumed even though only 1024 bytes have been
scanned

The scanner has an internal buffer size of `1024 bytes`. This PR copies
`scanner.Scanner` from the `text/scanner` package and increases the
internal buffer to `maxInputSize` which guarantees that no calls to
`Scan` will incur reads beyond the first one.

Ideally we wouldn't pass Scanners around in such a way but the cost of
that is to rethink much of Loki's lexing logic. Given we know a max
input size and we already have that in memory, this solution should
suffice.

Fixes https://github.com/grafana/loki/issues/9132