[helm] Bump GEL version, default to 3 scalable targets (#9170)
**What this PR does / why we need it**:

This PR bumps the GEL version to match the Loki version that was added
in version 5 of the helm chart. With these new versions we can now
default to 3 targets in scalable mode `read`, `write`, and `backend`.

This PR also fixes 2 bugs when running in scalable mode:

* the gateway should forward requests to the admin api to the `backend`
target when running 3 targets
* the index gateway should default to using `ring` mode, which will
allow the queriers in the new `read` target to find the index gateway
address for their index gateway clients via the ring.