jsonnet: remove mistakenly added deep merging
This was added in 4d55a24ed3cbed399c189dafb8c4140cbffce21d, probably to
align with the multizone ingesters below, but in this case the logic is
different, due to the super.ingester_statefulset in the line below. If
the merging is applied this way e.g. volumes are duplicated on the
statefulset.