[opt] s3: add s3 client http timeout for GetObject or PutObject (#8951)
**What this PR does / why we need it**:
add s3 client http timeout for GetObject or PutObject,

why？
The flush queue has been increasing. This is a difficult problem. We
have been investigating this problem for half a year. When the querier
has a large query, the load on the s3 server is too high and the number
of connections is full. It will cause the flush queue of the ingester to
increase continuously and cannot be reduced. When this problem occurs,
after we restart the s3 server, the flush queue will quickly become 0.

We checked with our s3 (object storage compatible with the s3 protocol)
team and found that it was because we used the s3 v1 sdk. The v1 sdk did
not configure timeout by default, and it would wait indefinitely without
configuration. The s3 v2 sdk defaults to a timeout of 30s, which only
causes write failures and does not cause loki to get stuck.

**Which issue(s) this PR fixes**:
Fixes #<issue number>
At present, as long as there are large logql queries such as 30 days, I
must restart the s3 server, otherwise our ingester queue will continue
to increase. We hope that we can break the stuck go routine after the 5m
timeout occurs.

**Special notes for your reviewer**:
ingester flush queue length up 🔝

![image](https://user-images.githubusercontent.com/9583245/228786401-cf07dc32-dcc0-428f-9558-561b588ea2c4.png)

big query, 30day logql or something happen

![image](https://user-images.githubusercontent.com/9583245/228789991-57185bc3-e628-45a5-82df-e4682590b786.png)

pprof goroutine list profile snapshot

![image](https://user-images.githubusercontent.com/9583245/228787732-6170b8ec-db86-4ca7-8ea0-a7f19b7d6141.png)

**There are other alternatives**:
```
ingester:
  flush_op_timeout: 10m
```
This `flush_op_timeout` configuration is more about the timeout of loki,
not the timeout of s3 client. We will also have various s3 operations
such as delete and getObject. We should add an independent timeout for
s3, just like the cassandra client has a separate The timeout period can
be configured.

And the parameter `flush_op_timeout` in the early version of loki
represents the timeout of writing `a batch of chunks `to s3, not a
timeout of s3. Now it is indeed changed to write to s3 one by one, but
it is not ruled out that in the future, it will be changed back to write
a batch of chunks to s3 sequentially. In this way, `flush_op_timeout:
5m` is far from enough to write a batch of chunks to s3.

```
  cassandra:
    timeout: 5m
```



**Checklist**
- [x] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [x] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`

---------

Co-authored-by: Michel Hollands <42814411+MichelHollands@users.noreply.github.com>