Docs: Fix API document
Remove /ready and /metrics from API list of the querier and the query
frontend because they are exposed by all components.

Remove /loki/api/v1/push from API list of the querier and the query
frontend because it is exposed by distributor.

Move /api/prom/push from API list of the querier and the query frontend
to API list of the distributor because it is exposed by distributor.

Add /api/prom/series to API list of the querier and the query frontend.

Change reference to /loki/api/v1/label to /loki/api/labels because
former is not documented.

Signed-off-by: Mitsuru Kariya <mitsuru.kariya@nttdata.com>