Bring back centralized docs toolkit (undo revert in #9146) (#9147)
**What this PR does / why we need it**:

This reverts the revert in https://github.com/grafana/loki/pull/9146
with a slightly less "nuclear" solution.

I think the problem was that CI was pulling `docs.mk` from `main`, and
it was creating a diff between the version we had checked in. I changed
`docs.mk` to point to a specific sha, which should prevent this in the
future.

I also opened a
[PR](https://github.com/grafana/writers-toolkit/pull/228) against the
writers-toolkit.