fix flaky integration TestMicroServicesDeleteRequest test (#9337)
**What this PR does / why we need it**:

`loki_compactor_deleted_lines` increments each time we find a line to
delete from a chunk. If the chunk is indexed in multiple tables, we
increment it for each of these tables as they get compacted.

Currently the chunk (that holds the line to be deleted) starts @`now -
46h`, so depending on when the test runs it gets indexed in 2 or 3
tables. It has been failing between hours `22:00 - 23:59` as the chunk
only overlaps and gets indexed in 2 tables while the test expects it to
be indexed in 3
https://drone.grafana.net/grafana/loki/23063/2/1
https://drone.grafana.net/grafana/loki/23067/2/1

Updating the chunk start time to now-48h so that it always gets indexes
in 3 tables