Add shuffle sharding to index gateway
This PR introduces shuffle sharding of the index files across index
gateways where the shuffle sharding key is the tenant ID.

Instead of having a global replication factor on the index gateways
which affects all tenants the same, whether they are small or large
tenants, the shuffle sharding allows to specify a shard factor per
tenant (on top of the replication factor).

The shuffle sharding algorithm is part of dskit's ring implementation
and is not part of this commit.

In case you have set a high RF for the index gateways, it is required to
reduce the RF and instead set the shard factor for the tenants.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>