Fix case insensitivity bug with additional checks
This change introduces two new boolean checks to the fast ascii case
when testing whether lower case values are contained in a filter.

The first checks that the character from substr is also a value below
the RuneSelf constant (to avoid incorrect matches such `d` = `„`).

The second check is added to the equality check for characters to allow
the substr character to be upper case and the line character to be
lower case. Previously this code required either both characters to be
the same case, or for the line to the upper and the substr to be lower.