Fix query behavior in Single Binary and SSD deployments (#9476)
To keep ingesters from calling themselves, the AsyncStore was only being
used when the module was `Querier`, `Ruler`, or `Read`. This means that,
for some operations, ingesters were never queries when the deployment
mode was single-binary or `Backend` (where the ruler now lives).

This change scopes the async store to just the querier rather than all
storage access and always includes in it queriers.

Because the `AsyncStore` isn't automatically instantiated, now, I've
also moved `IngesterQuerier` to be a dependency to the store. That way
anything that wants to make an `AsyncStore` is guaranteed to have what
it needs.