Helm: Quote tenantId value in logsInstance (#9486)
**What this PR does / why we need it**: Quotes the value of `tenantId`
for the cases where tenantId is made up of only numerical values (i.e.
if tenantId is `0`)

the motivation for this is that we have long used the tenant id `0` in
our loki instance for some stuff. We probably should fix that, but it's
going to be easier (at least for us) to patch this than to rename this
tenant or migrate to a new tenant.

**Which issue(s) this PR fixes**:
N/A - If you'd like me to file an issue to track this there as well, let
me know and i'll write one up. i figured this fix was simple enough

**Special notes for your reviewer**:

**Checklist**
- [x] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [x] Documentation added
- [x] Tests updated
- [x] `CHANGELOG.md` updated
- [x] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`