logql tests: use abstractions over Labels (#9533)
**What this PR does / why we need it**:

Call functions like `FromStrings`, so we don't have to know how the data
structure is laid out.

**Special notes for your reviewer**:

This PR only touches tests. 

Upstream Prometheus is using a different implementation of
`labels.Labels`, so I plan to create a series of PRs which make Loki
compatible with that. #9532 was the first.

**Checklist**
- [x] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- NA Documentation added
- [x] Tests updated
- NA `CHANGELOG.md` updated
- NA Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`
- NA For Helm chart changes bump the Helm chart version.

---------

Signed-off-by: Bryan Boreham <bjboreham@gmail.com>