Regex line/label filters return original regex when `String` is called (#9584)
To optimize regex line and label filters, Loki rewrites them to be
non-greedy. This can cause extra characters to appear in the regex when
`String()` is called. In addition to confusing logging, this can also
cause internal panics when a previously-parsable query becomes too long.

This PR stores the original regex in the label/line filter and uses that
in calls to `String`.

This PR also removes regex optimizations from label selectors. Making
the optimization at parse time caused the original regex to be lost.