Querier: configurable writeback queue bytes size (#9604)
**What this PR does / why we need it**:
The background writeback process is used to write chunks that have been
fetched from the store to cache. Chunks have a target size, but not all
chunks will necessarily be this size, so it is tricky to find the
correct value for `writeback_buffer` which specifies how many chunks
will be held in the queue at any given time.

For example: setting `writeback_buffer=10000` (the default) with all
chunks being the default target size (`chunk_target_size=1.5MiB`), in
the worst case could result in 15GiB of memory being used for this
buffer. Practically speaking, the queue never grows so deep, but if the
cache Loki is writing to becomes overwhelmed this could happen.

Setting a value too low would result in many chunks not being written
back to cache, which defeats the purpose.

This PR introduces a configurable _byte size_ limit to the writeback
cache, which defaults to 1GB in size. With this setting
`writeback_size_limit` in place, we get the best of both worlds: a deep
queue which can hold many small chunks or a shallow queue which can hold
a few large ones without using too much memory so as to overwhelm the
querier.

If either the `writeback_buffer` or `writeback_size_limit` reaches
capacity, any new chunks which are attempted to be added to the queue
will be dropped.

Two new metrics are added by this PR:

- `loki_cache_background_queue_bytes`: gauge, measures the size of the
queue in bytes
- `loki_cache_dropped_background_writes_bytes_total`: counter, measures
the volume of bytes dropped

(cherry picked from commit a248e3e97067b99e29511d9bd3285b4a789e5d20)