Jsonnet: add dns+ to join_members in memberlist config (#9723)
**What this PR does / why we need it**:

We use a headless service called `gossip-ring` which contains all the
memberlist members, previously we passed this headless service into our
memberlist code as

```
join_members:
  - gossip-ring.loki-ops.svc.cluster.local:7946
```

What this resulted in was the dskit wrapper around the memberlist
library thinking there was only 1 join_node, so it passed this directly
to the join method in memberlist. The memberlist lib would then do a DNS
resolution on this and it would find **all** the members behind this
service.

This is generally not noticeable until you get to very large clusters
with thousands of memberlist members and it results in new joiners
attempting to join to ALL the existing members instead of just a few.

This PR prefixes the url with `dns+` which changes the behavior inside
dskit in an important way.

```
join_members:
  - dns+gossip-ring.loki-ops.svc.cluster.local:7946
```

Now dskit will resolve the address to find all the memberlist members,
first, then apply code which selects a subset of all the nodes to be
used for first join (based on some logarithmic math)

This subset is then passed to the memberlist library which will join
just this set instead of all members.



**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Special notes for your reviewer**:

**Checklist**
- [ ] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`
- [ ] For Helm chart changes bump the Helm chart version in
`production/helm/loki/Chart.yaml` and update
`production/helm/loki/CHANGELOG.md` and
`production/helm/loki/README.md`. [Example
PR](https://github.com/grafana/loki/commit/d10549e3ece02120974929894ee333d07755d213)

Signed-off-by: Edward Welch <edward.welch@grafana.com>