Add shuffle sharding to index gateway (#9710) (#9752)
**Backport of #9710 to k155**

--- 

This PR introduces shuffle sharding of the index files across index
gateways where the shuffle sharding key is the tenant ID.

Instead of having a global replication factor on the index gateways
which affects all tenants the same, whether they are small or large
tenants, the shuffle sharding allows to specify a shard factor per
tenant (on top of the replication factor). This a) reduces the amount of
indexes that are downloaded on individual index gateways in case of a
high replication factor (which is currently the de-factor standard to
deal with high load) and b) allows to scale tenants individually.

The shuffle sharding algorithm is part of dskit's ring implementation
and is not part of this commit. If you have set a high RF for the index
gateways, it is required to reduce the RF and instead set the shard
factor for the tenants.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>