store: fix incorrect usage of period config in table range (#9754)
**What this PR does / why we need it**:
We are incorrectly taking reference to a loop variable in store init, I
overlooked this part as the reference was being taken in the call to
`GetIndexTableNumberRange`

As a result of this, `TableRange` of each period (which is used by
index-shipper) points to the latest `periodConfig`. This in itself is
not a problem if the index prefix doesn't change as we use the
`periodConfig` only to generate the `tableName`. But if the index prefix
changes across periods, we end up losing access to the indexes using the
old prefix.


**Special notes for your reviewer**:

**Checklist**
- [X] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [x] Tests updated
- [x] `CHANGELOG.md` updated
- [ ] If the change is worth mentioning in the release notes, add
`add-to-release-notes` label
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`
- [ ] For Helm chart changes bump the Helm chart version in
`production/helm/loki/Chart.yaml` and update
`production/helm/loki/CHANGELOG.md` and
`production/helm/loki/README.md`. [Example
PR](https://github.com/grafana/loki/commit/d10549e3ece02120974929894ee333d07755d213)

---------

Signed-off-by: Ashwanth Goli <iamashwanth@gmail.com>