Observe index gateway request count per tenant
This change adds a new `loki_index_gateway_client_requests_total` which
reports the total amount of requests performed by clients against the
index gateway.

Even though, there is already the histogram metric
`loki_index_gateway_request_duration_seconds`, which counts the
requests, the new metric also reports the tenant ID as well as the
status ("success" or "error"). The tenant ID cannot be reported by the
former metric, because it is implemented as part of the generic gRPC
client instrumentation.

The new metric allows to observe the RPS per tenant, so you can draw
conclusions about the required index gateway shard size per tenant.

Note that the new metric is only used when the index gateway is running
in ring mode.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>