fix integration test for multiple periods (#9782)
The integration test is writing logs across two index periods but
there's a chance that the log `lineB` is written after the beginning of
the second period after it's added. This means we never get `lineB` in
the results.

This change pushes back the write time of the first logs so they're
always stay in the first index period.