Migrate individual chunks as opposed to all chunks.   (#9780)
This way, if an error occurs, we get as much as we can.

**What this PR does / why we need it**:
In some situations, not all blob data may be accessible. For instance,
in the case where we have started migration between object storage
buckets, some data may be in the old bucket. If we wanted to migrate
data from the new bucket, those chunks will error with a 404, as they
are in the old bucket.

This PR is so that we continue on and get all of the chunks that we can,
rather than stopping with whatever error may arise.


**Checklist**
- [ ] Reviewed the
[`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md)
guide (**required**)
- [ ] Documentation added
- [ ] Tests updated
- [ ] `CHANGELOG.md` updated
- [ ] If the change is worth mentioning in the release notes, add
`add-to-release-notes` label
- [ ] Changes that require user attention or interaction to upgrade are
documented in `docs/sources/upgrading/_index.md`
- [ ] For Helm chart changes bump the Helm chart version in
`production/helm/loki/Chart.yaml` and update
`production/helm/loki/CHANGELOG.md` and
`production/helm/loki/README.md`. [Example
PR](https://github.com/grafana/loki/commit/d10549e3ece02120974929894ee333d07755d213)