Add per-tenant request counter metric to index gateway server
This commit add a counter metric with labels `operation`, `tenant`,
`status` for gRPC requests that are served by the index gateway.

What for?

The per-tenant RPS on the index gateway is used to derive the per-tenant
shard factor.

Why tracking on the server?

Unlike tracking index gateway RPS on the client side, tracking on the
server side does not yield that many series, even in multi-tenant
installations with a lot of tenants, because the amout of index gateway
instances is relatively small compared to the amount of queriers and
frontends.

Signed-off-by: Christian Haudum <christian.haudum@gmail.com>