index: allow refreshing index file names cache for specific tables (#9798)
**What this PR does / why we need it**:
We have a cache for index files available in the object storage, which
helps reduce the number of list calls we make to the object storage.
Currently, we only support refreshing the whole index cache, which
includes all the files for all tables and users. However, in a large
cell with many tables and tenants, it is less likely that all the tables
are being referenced. Also, most iterative operations would only need a
fresh listing of just a few tables at a time so it is wasteful to always
refresh the whole index cache.

This PR adds support for refreshing the index cache for only specific
tables. I am doing it as part of some compactor improvements to run
compaction and retention in separate goroutines to avoid blocking
compaction when some expensive delete requests are being processed. To
make those changes, I needed a way to cost-effectively refresh the index
files in the table being picked up for processing.

**Checklist**
- [x] Tests updated