Call FetchChunks in parallel from series index store (#9935)
This PR changes the invocation of `FetchChunks` in the series index
store to happen in parallel. It seems like this was the original
intention due to the `// bound concurrency` comment when allocating the
`groups` slice. This came out of an investigation of some slow queries,
and the traces showed many serial calls to `FetchChunks`.