operator: add initContainer to pods of zone aware components
Issue: https://issues.redhat.com/browse/LOG-3835

Problem: when a user configures zones in spec.replication we want to
make sure that when loki pods are started that they have access to the
zone that they are scheduled in. The zone_labeling_controller is
responsible for annotating the pods of zone aware componets, once these
are scheduled an annotation containing the zone where the pod was
scheduled is added. Furthermore, the controller will also configure the container
env vars to have this annotation available as an ENV variable.
Loki for loki to operate propperly this ENV var must be set when the
container starts, otherwise loki will not be configured propperly.

Solution: in this PR we have added an initContainer to each pod of zone
aware components. This initContainer will validate if the annotation was
set, by mounting the annotation as a volume and validating that the file
inside that volume is not empty
So if a container starts at T and the controller still hasn't added the
annotation then initContainer will enter a loop waiting for the
annotation to be added. Once the annotation is added since we use a volume
mount to access tha annotation the file inside the volume will be
updated, hence breaking the loop and allowing the loki components to
start.

Validation: To validate this solution I tested the PR on a AWS cluster
and validated the if the ENV is set for the loki processes with:

oc get pods -o json | jq --arg q "'\0' '\n'" -r ' .items[] | select( .metadata.name | test("distributor|index|ingester|querier|query")) | "echo \(.metadata.name) $(oc exec \(.metadata.name) -c \(.spec.containers[0].name) -- cat /proc/1/environ | tr \($q) | grep INSTANCE_AVAILIBILITY_ZONE)"' | bash